Extract long render function into RepositoryFileTreePage (#31923)
I ran into this while trying to change something in this render method
and thought this was better than one 100-line function
